Added QR codes
Added an api endpoint to generate an SVG QR code at /api/qrcode which is likely to change in the future. I should also consider changing the endpoint from a GET request to a POST request. I am also adding the SVG by editing the innerHTML of a div which is probably not ideal :p